Moves SOR bulkCreate unit tests to dedicated file
